+++ flask_acl/__init__.py
@@ -1,3 +1,3 @@
 from flask_acl.core import check
-from flask_acl.extension import AuthzManager
-from flask_acl.globals import current_authz
+from flask_acl.extension import ACLManager
+from flask_acl.globals import current_acl_manager