+++ flask_acl/globals.py
@@ -4,4 +4,4 @@ import werkzeug as wz
 from flask import current_app
 
 # Proxy to the current app's AuthManager
-current_authz = wz.local.LocalProxy(lambda: current_app.authz_manager)
+current_acl_manager = wz.local.LocalProxy(lambda: current_app.acl_manager)